feat : 팀 스코어 업데이트 로직

+++ src/main/java/com/moing/backend/domain/missionState/domain/service/MissionStateQueryService.java
@@ -13,7 +13,7 @@ public class MissionStateQueryService {
 
     private final MissionStateRepository missionStateRepository;
 
-    public Long stateCountByMissionId(Long missionId) {
+    public int stateCountByMissionId(Long missionId) {
         return missionStateRepository.getCountsByMissionId(missionId);
     }
